pr feedback: rename shared -> sharedcontext, remove transition from split button dropdown
