+++ ui/apps/dashboard/src/app/SharedContextProvider.tsx
@@ -1,7 +1,10 @@
 'use client';
 
-import { SharedProvider, type SharedHandlers } from '@inngest/components/Shared/SharedContext';
-import { legacyTraceToggle } from '@inngest/components/Shared/useLegacyTrace';
+import {
+  SharedProvider,
+  type SharedHandlers,
+} from '@inngest/components/SharedContext/SharedContext';
+import { legacyTraceToggle } from '@inngest/components/SharedContext/useLegacyTrace';
 
 import { useInvokeRun } from '@/queries/useInvokeRun';
 import { useRerun } from '@/queries/useRerun';
