+++ ui/apps/dashboard/src/queries/useInvokeRun.tsx
@@ -1,4 +1,4 @@
-import type { InvokeRunPayload } from '@inngest/components/Shared/useInvokeRun';
+import type { InvokeRunPayload } from '@inngest/components/SharedContext/useInvokeRun';
 import { useMutation } from 'urql';
 
 import { useEnvironment } from '@/components/Environments/environment-context';
