+++ ui/apps/dev-server-ui/src/app/(dashboard)/run/page.tsx
@@ -2,7 +2,7 @@
 
 import { RunDetailsV2 } from '@inngest/components/RunDetailsV2/RunDetailsV2';
 import { RunDetailsV3 } from '@inngest/components/RunDetailsV3/RunDetailsV3';
-import { useLegacyTrace } from '@inngest/components/Shared/useLegacyTrace';
+import { useLegacyTrace } from '@inngest/components/SharedContext/useLegacyTrace';
 import { useSearchParam } from '@inngest/components/hooks/useSearchParam';
 import { cn } from '@inngest/components/utils/classNames';
