+++ ui/packages/components/src/RunDetailsV3/RerunModal.tsx
@@ -1,11 +1,11 @@
 import { useState } from 'react';
-import { RiCheckLine, RiCheckboxCircleFill, RiExternalLinkLine } from '@remixicon/react';
+import { RiCheckboxCircleFill, RiExternalLinkLine } from '@remixicon/react';
 import { toast } from 'sonner';
 
 import { Alert } from '../Alert';
 import { Link } from '../Link';
 import { AlertModal } from '../Modal';
-import { useRerun } from '../Shared/useRerun';
+import { useRerun } from '../SharedContext/useRerun';
 
 type RerunProps = {
   runID: string;
